requirements.txt
pandas
kafka-python
six
boto3
s3fs
kafka-python-ng